pyproject.toml
[project]
name = "qut-data-analytics-duckdb-final"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.0"
dependencies = [
    "dbt-bigquery>=1.10.2",
    "dbt-core>=1.10.11",
    "gcloud>=0.18.3",
    "google-auth>=2.40.3",
    "google-auth-oauthlib>=1.2.2",
    "google-cloud-bigquery>=3.37.0",
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "plotly-express>=0.4.1",
    "streamlit>=1.49.1",
]
